recovery: change delivery_rate to u64
delivery_rate is defined as bytes/sec, so no need to be f64.

+++ src/ffi.rs
@@ -636,7 +636,7 @@ pub extern fn quiche_conn_stats(conn: &Connection, out: &mut Stats) {
     out.lost = stats.lost;
     out.rtt = stats.rtt.as_nanos() as u64;
     out.cwnd = stats.cwnd;
-    out.delivery_rate = stats.delivery_rate as u64;
+    out.delivery_rate = stats.delivery_rate;
 }
 
 #[no_mangle]
